renamed process.py to helpers.py moved function keyword_check to helpers.py
